add warnings tests
